Staging biharmonic for GPU porting
The answer is still in line with CPU flag changes. In this commit, I've lined up the loop nests so that I can collapse them entirely (except for limiter_optim_iter_full at this point), but all of the pre-computations of qtens_biharmonic are structured in such a way that they can be ported using OpenACC efficiently. Some side tests on how to perform the reductions of qmin and qmax demonstrated that the most efficient way to do this is to ignore OpenACC's reduction clauses, collapse the loop nest entirely, and simply perform the reduction manually using global data structures.