Partial OpenACC port of qtens_biharmonic
This includes several changes:

(1) Removed most of the data transfers and still get the right answer
(2) Includes copy of qdp1 after time_avg routine if limiter_option==8
(3) Alters threading of euler_hypervis_kernel_last to loop over fewer than nlev levels
(4) Ports a few of the reduction routines to GPU using OpenACC with poor data transfers

It also discovered a bug in OpenACC for PGI at least. The last kernel ported to OpenACC had to be commented out here because it is giving the wrong answer on the GPU. Technically, this should never be the case because if it works on the CPU, automagical data copies should ensure it works on the GPU. I need to find out what's going on and how to work around it. Perhaps I just need to try 15.4.