Fixes for OpenACC (no OACC bug in last commit)
Continuing to port the biharmonic and reduction computations for OpenACC. Couple important findings / reminders
(1) Need to put "gang vector" on the loop directives
(2) There was no OpenACC bug in the last commit. It was a race condition.
(3) minval and maxval are not ported to OpenACC yet.

I'm getting the right answer with what's ported so far.