Biharmonic staged for OACC porting
Nearly all of the limiter_option==8 material is staged for OpenACC porting. Last time around, there was a threading problem in edgeVpack, in which when you loop over 1,nelemd with master, you get wrong answers. With my local pack/unpack routines, this is no longer the case, thought it's not entirely clear why. Every portion of limiter_option == 8 has been moved to loop over 1,nelemd with the master thread. I've tested everything to be machine precision the same with Cray versus PGI.