PGI bug workaround; now OpenACC giving wrong ans
I worked around the PGI bug from the previous commit, noticing that they were creating summation reductions for addmass in limiter_optim_iter_full, and then changing that code so that the compiler could not pick up on a reduction-like situation (which it shouldn't have been doing anyway since that loop isn't partitioned over threads). Now, however, the OpenACC code for limiter_optim_iter_full is giving the wrong answer, even though the CPU code is giving the right answer. I'm not sure yet if this is a bug in the compiler of an oversight in my code (race conditions, etc). However, in the meantime, the next commit will take Irina's CUDA implementation of limiter_optim_iter_full and use that instead until the OpenACC stuff is worked out.