Lim8 OpenACC correct & more PCI-e removed
Forgot I had some PCI-e traffic with Qdp in there from the initial porting effort. That has been removed. Also, I had to add a copy of the first tracer to the host in qdp_time_avg only when limiter_option == 8 to give correctness for the first tracer. Now, the code is correct to machine precision for limiter_option == 8 and limiter_option == 4.