Further optimizataions, lim8 working on GPU
All answers machine precision the same. I added vector_length clauses to some kernels, and I replaced large array accesses with temp variables in the most expensive. The limiter_optim_iter_full kernel, in particular, is an order of magnitude cheaper now. The limiter_option == 8 port, as it stands is 2.4x faster in the tracer routines and 1.9x faster for the HOMME dycore when using 64 elements per node. I'll look for ways to improve this in the future. The kernel timings all seem very high by a factor of 10 in many places, and I need to find out how to improve them perhaps via shared memory or some other optimization.