Beginning shared memory with OpenACC
I've started using shared memory in a few kernels with OpenACC. They look very ugly because I've had to manually fision and fuse loops, but the speed-up is good. I need to finish altering the unpacks, and I need to change laplace_sphere_wk so that it it works on just on i,j at a time, and then refactor the loop so that lap_p is in shared memory. After that, the code should have quite good speed-up with limiter8.