Add glc_import and _export routines, set glc_cpl_indices
This commit overhauls the glc_import, glc_export, and glc_cpl_indices
routines so that they are set up properly for MPASLI.  Any previous code
in those routines was non-functional remnants of versions from CISM and
MPASO and were based on the pre-CIME coupler.  The new version is based
on CIME and therefore has a proper GLC grid and remapping and
downscaling of SMB in the coupler (GLINT layer eliminated).

Only coupler fields related to SMB are addressed in this commit, but the
framework should easily be extensible to other coupling fields.

With this commit, visually correct SMB fields are coming from the
coupler on to the MPASLI grid, though there may be some additional issues
to sort out.