Add glc_export routine and set glc_cpl_indices
This commit adds the contents of a glc_export routine to be consistent
with the coupler in CIME.

It also correctly sets the glc_cpl_indices values which, though already
existed, were a mishmash of things.

With this commit, obvious SMB fields are coming from the coupler on to the
MPASLI grid, though there may be some additional issues to sort out.