Add calculation of glc->cpl masks
GLC needs to send the coupler 3 different masks:
ice_covered, ice_sheet_grid_mask, icemask_coupled_fluxes
This commit adds the calculations for the last two.  (The first was
already added.)  Note that an assumption is made that MPASLI will always
be evolving.  This could be relaxed later with additional changes, if
required.

It also adds descriptions of all the GLC coupling fields in comments in
glc_cpl_indices.F.

With these changes SMB coupling should be working correctly.