Update fglcmask in clm45 namelist defaults
The CLM namelist includes paths to fglcmask files which indicate where
SMB should be calculated within CLM.  This commits a change that will
always use a mask that covers all of Greenland and Antarctica.  (This
will be somewhat inefficient when only one of the ice sheets is
required.)

There is logic in namelist_defaults that adjusts the mask file to be
used based on the CLM grid being used and GLC being used.  However,
the logic for the GLC grid does not appear to be working.  Therefore,
this composite AIS/GIS mask enables running ACME for either ice sheet
without requiring complicated changes.  When that logic is fixed, the
choice of mask files will need to be updated.  I have added some
comments about how to do that.