Update config_compilers for linking Albany on LCFs.
Figured out how to link Albany to ACME on Titan and Mira. This
required selection of a linker and additional libs at link time.

On Titan, using pgi for fortran and gcc for Albany, and fortran linker.
On Mira, using xlf for fortran, bgclang++ for Albany, and C++ linker.

On Mira, there was an issue with name mangling. Added a new define
on Mira (MPASLI_EXTERNAL_INTERFACE_DISABLE_MANGLING) to turn off
name mangling in Interface_Velocity_External.hpp. Needs change
in mpasli submodule to match.

One change needs to be checked: Mira/ibm used to add -qsmp=omp
as LDFLAG. I had to remove this with C++ linking since it is
not recognized by the C++ compiler. We need to verify that this
doesn't mess up compiling or performance for other components.
This flag is still there for Fortran compiles, just not linking.

[BFB]