Fixes bugs in energy conservation fixer (CLUBB) and restart physics
This PR fixes bugs reported by NCAR in CLUBB and restart physics.
Following is their description:
CLUBB:
=====
Modify the energy conservation fixer in the CAM CLUBB interface to
remove small tendencies CLUBB may produce in layers above which it
is active to the layers where it is active, to avoid dehydration in
the stratosphere. This also includes a minor bug fix to the
pre-existing energy fixer.
-->M components/cam/src/physics/cam/clubb_intr.F90
modifications to energy conservation fixing
-->M components/cam/src/physics/cam/macrop_driver.F90
minor bug fix for ice supersaturation
This was implemented in cam5_4_37

Restart Physics:
===============

Jim Edwards discovered a bug and provided a fix in restart_physics.
This bug only occurs when using pio2 (which is not yet the CAM default)
 and using the SE dycore.
-->M components/cam/src/physics/cam/restart_physics.F90
bug fix when using pio2
This went into: cam5_4_32

[BFB]??

AG-411, AG-413