Convenient way to run OpenACC from create_newcase
I put a new option flag in create_newcase called "-camse_target", which takes one option that specifies the HOMME cmake target to use. It defaults to 'preqx' if the flag isn't used. This option does two things in the create_newcase script:

(1) It sets CAM_TARGET in env_build.xml
(2) It set the default compiler to "pgi_acc" only if -mach == titan. Otherwise does not change default compiler. Also, this is done in such a way that if -compiler is specified, then it will overwrite pgi_acc.

So, to run OpenACC code on Titan, one simply needs to say:

./create_newcase -camse_target preqx_acc

The current valid values for camse_target are 'preqx' and 'preqx_acc'