Switch working branch to v0.3 version of the code.
*********1*********2*********3*********4*********5*********6*********7**
This commit serves as the baseline for the Process Ordering work on ACME
v0.3.  As a result, the v1 work brought in from the master branch has
been deleted and the v0.3 code has been moved into the directory.
	* Deleted body of code for v1
	* Copied in v0.3 of the ACME code

A good commit message should be written like an email, a subject
followed by a blank line, followed by a more descriptive body.

[BFB] - Bit-For-Bit