Removing unused variables from HOMME.
*********1*********2*********3*********4*********5*********6*********7**
Compiled with ifort on skybridge with flag '-warn unused'. There are screens of warnings of unused variables in homme. This is an attempt to clean some of them from files listed below.

Variables in #ifdefs and variables in some disabled code for debugging or reference were not cleared. In more obvious cases when the variable is really never used, the variable is deleted.

Should be:
[BFB] - Bit-For-Bit

So far
	modified:   coordinate_systems_mod.F90
	modified:   derivative_mod_base.F90
	modified:   edge_mod_base.F90
	modified:   gridgraph_mod.F90
	modified:   ll_mod.F90
	modified:   metagraph_mod.F90
	modified:   viscosity_mod_base.F90