Added flag to move convective detrainment into convective subroutines.
*********1*********2*********3*********4*********5*********6*********7**
Added a flag to handle convective detrainment after convection (deep or
shallow) calls instead of macrophysics.

Convective detrainment is usually handled within the macrophysics call,
macrop_driver.F90.  This can have the effect of making it seem like
only macrophysics contributes to liquid and ice levels.  A namelist
variable, detrain_conv, has been added that allows for the shift of
handling convective detrainment after each of the respective convection
calls (deep and shallow) instead of within macrophysics.
	* namelist variable 'detrain_conv', if made .true., will cause
          convective detrainment to be calculated following the
	  convection processes calls (either deep or shallow), instead
	  of within the macrophysics call.
	* fixed a bug with cmfmc variable within the physpkg.F90 code
	  where only the deep convection contribution was passed to
          macrophysics and microphysics.

[Non-BFB] - Non Bit-For-Bit

Note: If the handling of detrainment is moved this will change the
water content in the model state before certain physics processes are
called.  This is therefore not bit-for-bit.

See confluence for a more detailed description about these tags.